+++ controllers/apps/cluster/cluster_plan_builder.go
@@ -60,7 +60,8 @@ type clusterTransformContext struct {
 	components []*appsv1.ClusterComponentSpec
 	shardings  []*appsv1.ClusterSharding
 
-	shardingComps map[string][]*appsv1.ClusterComponentSpec // comp specs for each sharding
+	shardingComps        map[string][]*appsv1.ClusterComponentSpec // comp specs for each sharding
+	shardingCompsWithTpl map[string]map[string][]*appsv1.ClusterComponentSpec
 
 	// TODO: remove this, annotations to be added to components for sharding, mapping with @allComps.
 	annotations map[string]map[string]string
