+++ controllers/apps/cluster/transformer_cluster_component_status_test.go
@@ -59,7 +59,7 @@ var _ = Describe("cluster component status transformer", func() {
 		transCtx.components, transCtx.shardings, err = transformer.resolveCompsNShardingsFromSpecified(transCtx, cluster)
 		Expect(err).Should(BeNil())
 
-		transCtx.shardingComps, err = transformer.buildShardingComps(transCtx)
+		transCtx.shardingComps, _, err = transformer.buildShardingComps(transCtx)
 		Expect(err).Should(BeNil())
 	}
